+++ src/main/java/dao/ChessGameDao.java
@@ -1,19 +1,15 @@
 package dao;
 
 import chess.ChessGame;
-import java.util.List;
+import chess.domain.piece.Piece;
+import chess.domain.position.Position;
+import java.util.Map;
 
 public interface ChessGameDao {
 
-    List<String> gameIds();
-
     void save(ChessGame chessGame, final String gameId);
 
-    String createChessStatus(final ChessGame chessGame);
-
-    ChessGame select(final String gameId);
-
-    void update(ChessGame chessGame, final String gameId);
+    Map<Position, Piece> select(final String gameId);
 
     void reset(final String gameId);
 }
